beyond: fix the debug-symbol checkout
shutil.rmtree makes an error when called on an non existing tree.

+++ beyond/src/__init__.py
@@ -58,12 +58,13 @@ def symbols_update():
   if symbols_checkout():
     return True
   # We failed.  Maybe the repository is broken, erase it and restart.
-  try:
-    shutil.rmtree(repo_dir)
-  except Exception as e:
-    log('symbolize: symbols_update: cannot remove {}: {}',
-        repo_dir, e)
-    return False
+  if os.path.exists(repo_dir):
+    try:
+      shutil.rmtree(repo_dir)
+    except Exception as e:
+      log('symbolize: symbols_update: cannot remove {}: {}',
+          repo_dir, e)
+      return False
   p = run(['git', 'clone',
            'git@git.infinit.sh:infinit/debug-symbols.git',
            repo_dir])
